js-decompiler/pyproject.toml
[project]
name = "js-decompiler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.13.4",
    "fastapi==0.116.1",
    "langchain==0.3.27",
    "langchain-core==0.3.72",
    "langchain-google-genai==2.1.8",
    "langchain-openai==0.3.28",
    "outlines>=1.2.4",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    "streamlit==1.36.0",
]
